+++ parser/src/test/java/io/github/cowwoc/styler/parser/test/IntegrationTest.java
@@ -1,6 +1,7 @@
 package io.github.cowwoc.styler.parser.test;
 
 import io.github.cowwoc.styler.ast.core.NodeIndex;
+import io.github.cowwoc.styler.parser.ParseResult;
 import io.github.cowwoc.styler.parser.Parser;
 import org.testng.annotations.Test;
 
@@ -37,7 +38,9 @@ public class IntegrationTest
 
 		try (Parser parser = new Parser(source))
 		{
-			NodeIndex root = parser.parse();
+			ParseResult result = parser.parse();
+			requireThat(result, "result").isInstanceOf(ParseResult.Success.class);
+			NodeIndex root = ((ParseResult.Success) result).rootNode();
 			requireThat(root.isValid(), "root.isValid()").isTrue();
 		}
 	}
@@ -58,7 +61,9 @@ public class IntegrationTest
 
 		try (Parser parser = new Parser(source))
 		{
-			NodeIndex root = parser.parse();
+			ParseResult result = parser.parse();
+			requireThat(result, "result").isInstanceOf(ParseResult.Success.class);
+			NodeIndex root = ((ParseResult.Success) result).rootNode();
 			requireThat(root.isValid(), "root.isValid()").isTrue();
 		}
 	}
@@ -79,7 +84,9 @@ public class IntegrationTest
 
 		try (Parser parser = new Parser(source))
 		{
-			NodeIndex root = parser.parse();
+			ParseResult result = parser.parse();
+			requireThat(result, "result").isInstanceOf(ParseResult.Success.class);
+			NodeIndex root = ((ParseResult.Success) result).rootNode();
 			requireThat(root.isValid(), "root.isValid()").isTrue();
 		}
 	}
@@ -100,7 +107,9 @@ public class IntegrationTest
 
 		try (Parser parser = new Parser(source))
 		{
-			NodeIndex root = parser.parse();
+			ParseResult result = parser.parse();
+			requireThat(result, "result").isInstanceOf(ParseResult.Success.class);
+			NodeIndex root = ((ParseResult.Success) result).rootNode();
 			requireThat(root.isValid(), "root.isValid()").isTrue();
 		}
 	}
@@ -120,7 +129,9 @@ public class IntegrationTest
 
 		try (Parser parser = new Parser(source))
 		{
-			NodeIndex root = parser.parse();
+			ParseResult result = parser.parse();
+			requireThat(result, "result").isInstanceOf(ParseResult.Success.class);
+			NodeIndex root = ((ParseResult.Success) result).rootNode();
 			requireThat(root.isValid(), "root.isValid()").isTrue();
 		}
 	}
@@ -142,7 +153,9 @@ public class IntegrationTest
 
 		try (Parser parser = new Parser(source))
 		{
-			NodeIndex root = parser.parse();
+			ParseResult result = parser.parse();
+			requireThat(result, "result").isInstanceOf(ParseResult.Success.class);
+			NodeIndex root = ((ParseResult.Success) result).rootNode();
 			requireThat(root.isValid(), "root.isValid()").isTrue();
 		}
 	}
@@ -164,7 +177,9 @@ public class IntegrationTest
 
 		try (Parser parser = new Parser(source))
 		{
-			NodeIndex root = parser.parse();
+			ParseResult result = parser.parse();
+			requireThat(result, "result").isInstanceOf(ParseResult.Success.class);
+			NodeIndex root = ((ParseResult.Success) result).rootNode();
 			requireThat(root.isValid(), "root.isValid()").isTrue();
 		}
 	}